Added user permission

+++ app/Config/Routes.php
@@ -5,40 +5,48 @@ use CodeIgniter\Router\RouteCollection;
 /**
  * @var RouteCollection $routes
  */
-$routes->get('/', 'Home::index');
+$routes->get('/', 'Home::index', ['filter' => 'role:user,admin,manajer,direktur']);
 
 // Akun 1
-$routes->get('/akun1', 'Akun1::index');
-$routes->get('/akun1/new', 'Akun1::new');
-$routes->get('/akun1/edit/(:any)', 'Akun1::edit/$1');
-$routes->post('/akun1', 'Akun1::store');
-$routes->put('/akun1/edit/(:any)', 'Akun1::update/$1');
-$routes->delete('/akun1/(:any)', 'Akun1::destroy/$1');
+$routes->get('/akun1', 'Akun1::index', ['filter' => 'role:admin']);
+$routes->get('/akun1/index', 'Akun1::index', ['filter' => 'role:admin']);
+$routes->get('/akun1/new', 'Akun1::new', ['filter' => 'role:admin']);
+$routes->get('/akun1/edit/(:any)', 'Akun1::edit/$1', ['filter' => 'role:admin']);
+$routes->post('/akun1', 'Akun1::store', ['filter' => 'role:admin']);
+$routes->put('/akun1/edit/(:any)', 'Akun1::update/$1', ['filter' => 'role:admin']);
+$routes->delete('/akun1/(:any)', 'Akun1::destroy/$1', ['filter' => 'role:admin']);
 
 // Akun 2
-$routes->resource('akun2');
+$routes->resource('akun2', ['filter' => 'role:admin']);
 
 // Transaksi
-$routes->get('/transaksi/akun2', 'Transaksi::akun2');
-$routes->get('/transaksi/status', 'Transaksi::status');
-$routes->resource('transaksi');
+$routes->get('/transaksi/akun2', 'Transaksi::akun2', ['filter' => 'role:admin']);
+$routes->get('/transaksi/status', 'Transaksi::status', ['filter' => 'role:admin']);
+$routes->resource('transaksi', ['filter' => 'role:admin']);
 
 // Jurnal Umum
-$routes->get('/jurnalumum', 'JurnalUmum::index');
-$routes->post('/jurnalumum', 'JurnalUmum::index');
-$routes->post('/jurnalumum/printjurnalumum', 'JurnalUmum::printjurnalumum');
+$routes->get('/jurnalumum', 'JurnalUmum::index', ['filter' => 'role:user,admin']);
+$routes->get('/jurnalumum/index', 'JurnalUmum::index', ['filter' => 'role:user,admin']);
+$routes->post('/jurnalumum', 'JurnalUmum::index', ['filter' => 'role:user,admin']);
+$routes->post('/jurnalumum/printjurnalumum', 'JurnalUmum::printjurnalumum', ['filter' => 'role:admin']);
 
 // Posting
-$routes->get('/posting', 'Posting::index');
-$routes->post('/posting', 'Posting::index');
-$routes->post('/posting/printposting', 'Posting::printposting');
+$routes->get('/posting', 'Posting::index', ['filter' => 'role:admin']);
+$routes->get('/posting/index', 'Posting::index', ['filter' => 'role:admin']);
+$routes->post('/posting', 'Posting::index', ['filter' => 'role:admin']);
+$routes->post('/posting/printposting', 'Posting::printposting', ['filter' => 'role:admin']);
 
 // Neraca Saldo
-$routes->get('/neracasaldo', 'NeracaSaldo::index');
-$routes->post('/neracasaldo', 'NeracaSaldo::index');
-$routes->post('/neracasaldo/printneracasaldo', 'NeracaSaldo::printneracasaldo');
+$routes->get('/neracasaldo', 'NeracaSaldo::index', ['filter' => 'role:direktur']);
+$routes->get('/neracasaldo/index', 'NeracaSaldo::index', ['filter' => 'role:direktur']);
+$routes->post('/neracasaldo', 'NeracaSaldo::index', ['filter' => 'role:direktur']);
+$routes->post('/neracasaldo/printneracasaldo', 'NeracaSaldo::printneracasaldo', ['filter' => 'role:direktur']);
 
 // Laba Rugi
-$routes->get('/labarugi', 'LabaRugi::index');
-$routes->post('/labarugi', 'LabaRugi::index');
-$routes->post('/labarugi/printlabarugi', 'LabaRugi::printlabarugi');
+$routes->get('/labarugi', 'LabaRugi::index', ['filter' => 'role:direktur']);
+$routes->get('/labarugi/index', 'LabaRugi::index', ['filter' => 'role:direktur']);
+$routes->post('/labarugi', 'LabaRugi::index', ['filter' => 'role:direktur']);
+$routes->post('/labarugi/printlabarugi', 'LabaRugi::printlabarugi', ['filter' => 'role:direktur']);
+
+// User
+$routes->resource('user', ['filter' => 'role:manajer']);
